server/requirements.txt
fastapi
SQLAlchemy
psycopg2
bcrypt
PyJWT
uvicorn
python-dotenv
minio
python-multipart
payos==1.1.0
httpx
pytest
firebase-admin
